This PR replaces `ConcurrentExecPool` as a background thread inside `QueueGenericThreadsNonBlockingImpl` by the much simpler `CallbackThread`. Furthermore, since it is no longer necessary, the detaching logic in `ConcurrentExecPool` is removed. The implementation for `Enqueue` for `QueueUniformCudaHipRt` is also vastly simplified.

This PR causes one change in behavior: the destructor of a non-blocking queue will now block until all tasks in the queue are completed. This is required to ensure deterministic lifetime and destruction of queues, background threads, tasks, etc. If we are fine with this new behavior, we should also cover it by a unit test.

This changeset fixes several issues reported by TSAN (see #1936), although it does not completely fix all errors. I am suspecting there could be some remaining false positives.

Depends on:

- [x] Merge #1975 first
- [x] Merge #1987 first

When this PR is merged, `ConcurrentExecPool` is only used in one place anymore, as a yielding thread pool for running CPU kernels of the Threads backend. I think we could cut down the class to it's new purpose and potentially optimize performance.