@fwyzard 2023-06-24
Is there some benchmark that can show the impact on the runtime performance ?