@psychocoderHPC 2023-06-26
@bernhardmgruber based on this PR I added a few changes on top to silence all errors, I do not understand why some changes are required and would like to understand it.

https://github.com/bernhardmgruber/alpaka/pull/1