@bernhardmgruber 2023-06-28
I found the cause of the ASAN errors. There was a cyclic reference of shared pointers in some event implementations of the form: shared_ptr -> eventImpl -> std::future -> task lambda -> captured shared_ptr. This created a leak that ASAN diagnosed.