@bernhardmgruber 2023-06-28
> Is there some benchmark that can show the impact on the runtime performance ?

I added one testing 1000 enqueues.

Before:
```
-------------------------------------------------------------------------------
enqueueBenchmark - alpaka::test::TestQueues - 0
-------------------------------------------------------------------------------
/home/bgruber/dev/alpaka/test/unit/queue/src/QueueTest.cpp:201
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Enqueue 1000                                   100             4     2.0908 ms 
                                         5.1214 us     5.1144 us    5.13593 us 
                                        49.4037 ns    28.2211 ns    98.2177 ns 
                                                                               

-------------------------------------------------------------------------------
enqueueBenchmark - alpaka::test::TestQueues - 1
-------------------------------------------------------------------------------
/home/bgruber/dev/alpaka/test/unit/queue/src/QueueTest.cpp:201
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Enqueue 1000                                   100             1    78.2606 ms 
                                        961.536 us    953.695 us    966.312 us 
                                         30.501 us     21.397 us    53.1851 us 
                                                                               

-------------------------------------------------------------------------------
enqueueBenchmark - alpaka::test::TestQueues - 2
-------------------------------------------------------------------------------
/home/bgruber/dev/alpaka/test/unit/queue/src/QueueTest.cpp:201
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Enqueue 1000                                   100             1     1.21467 s 
                                        12.8079 ms    11.8978 ms     13.708 ms 
                                        4.61803 ms    4.24741 ms    5.02597 ms 
                                                                               

-------------------------------------------------------------------------------
enqueueBenchmark - alpaka::test::TestQueues - 3
-------------------------------------------------------------------------------
/home/bgruber/dev/alpaka/test/unit/queue/src/QueueTest.cpp:201
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Enqueue 1000                                   100             1    806.096 ms 
                                        8.02204 ms    7.98338 ms    8.05383 ms 
                                        178.694 us    151.852 us    202.262 us
```

After:
```
-------------------------------------------------------------------------------
enqueueBenchmark - alpaka::test::TestQueues - 0
-------------------------------------------------------------------------------
/home/bgruber/dev/alpaka/test/unit/queue/src/QueueTest.cpp:201
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Enqueue 1000                                   100             4     2.1004 ms 
                                        5.05373 us    5.04876 us    5.06079 us 
                                        29.8689 ns    20.7356 ns    39.9793 ns 
                                                                               

-------------------------------------------------------------------------------
enqueueBenchmark - alpaka::test::TestQueues - 1
-------------------------------------------------------------------------------
/home/bgruber/dev/alpaka/test/unit/queue/src/QueueTest.cpp:201
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Enqueue 1000                                   100             1     21.604 ms 
                                        299.072 us    293.559 us    302.797 us 
                                        22.8706 us    16.7133 us    29.3113 us 
                                                                               

-------------------------------------------------------------------------------
enqueueBenchmark - alpaka::test::TestQueues - 2
-------------------------------------------------------------------------------
/home/bgruber/dev/alpaka/test/unit/queue/src/QueueTest.cpp:201
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Enqueue 1000                                   100             1     1.18284 s 
                                        11.5493 ms    11.5286 ms      11.57 ms 
                                        105.008 us    85.5805 us    137.062 us 
                                                                               

-------------------------------------------------------------------------------
enqueueBenchmark - alpaka::test::TestQueues - 3
-------------------------------------------------------------------------------
/home/bgruber/dev/alpaka/test/unit/queue/src/QueueTest.cpp:201
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Enqueue 1000                                   100             1    881.351 ms 
                                        11.0215 ms    11.0092 ms    11.0346 ms 
                                        64.5709 us    56.8905 us    77.8324 us 
```
Queue order of `alpaka::test::TestQueues` is: CPU blocking, CPU non-blocking, CUDA blocking, CUDA non-blocking.

Perf got faster for all queues except for the CUDA non-blocking.