@bernhardmgruber 2023-06-28
> Any idea why CUDA async is somewhat slower ?

I have no clue. `perf` shows less than 1% of the code is spent in the test executable. Most time is spent inside Catch2 and the CUDA libraries. A potential guess could be that previously, we launched the CUDA host func and enqueued to the background thread at the same time, so some kind of thread switching overhead could overlap. Now, I enqueue a host func to CUDA, which then enqueues to the callback thread. The thread switching does not overlap in this case.

Also the numbers seem to be very shaky, even though Catch2 already reports the mean of a 100 runs. Running again a couple of times, and only the async CUDA queue, now shows:

Before:
```
-------------------------------------------------------------------------------
enqueueBenchmark - alpaka::test::TestQueues - 3
-------------------------------------------------------------------------------
/home/bgruber/dev/alpaka/test/unit/queue/src/QueueTest.cpp:201
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Enqueue 1000                                   100             1    785.955 ms 
                                        10.7045 ms    10.6835 ms    10.7233 ms 
                                        100.935 us    76.4397 us    153.905 us 
```

After, 1st run:
```
-------------------------------------------------------------------------------
enqueueBenchmark - alpaka::test::TestQueues - 3
-------------------------------------------------------------------------------
/home/bgruber/dev/alpaka/test/unit/queue/src/QueueTest.cpp:201
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Enqueue 1000                                   100             1    753.501 ms 
                                        11.2029 ms    11.1918 ms    11.2191 ms 
                                        67.5108 us    49.7149 us    93.5378 us 
```

After, 2nd run:
```
-------------------------------------------------------------------------------
enqueueBenchmark - alpaka::test::TestQueues - 3
-------------------------------------------------------------------------------
/home/bgruber/dev/alpaka/test/unit/queue/src/QueueTest.cpp:201
...............................................................................

benchmark name                       samples       iterations    estimated
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Enqueue 1000                                   100             1     524.13 ms 
                                        6.62121 ms    6.27566 ms     7.0501 ms 
                                        1.96068 ms     1.6434 ms    2.24793 ms 

```
The numbers are from my workstation. I think for a solid number one would need to run on a cluster.

> That's one of the cases we do care about :)

The entire changeset here only affects the implementation of host functions and does not affect memory copies or kernels. Are host functions really such a concern?

Also: this PR fixes several race conditions during the shutdown of alpaka and introduces consistent destruction and cleanup of resources as well as ensures all enqueued work is still finished. That is also a huge plus IMO.